Use INPUT_PULLUP
Previously, the mode was configured to INPUT and then written HIGH,
which works on AVR but isn't very portable. Instead, this uses the
INPUT_PULLUP pin mode (available since Arduino 1.0.1) which is portable.

Additionally, instead of driving the pin high for 40uS, the pullup is
just enabled again, letting that pull the pin high again (this should
also prevent problems if the sensor wants to drive the pin low while the
Arduino is still driving it high). This was tested using a DHT11 and
DHT22.