Let read() use pulseIn()
Previously, an inexact timing loop was used, where a loop count was
passed in at initialization time because the duration of a single loop
dependend on all kinds of factors (compiler, CPU speed, architecture).

Using micros() or delayMicroSeconds() for this kind of timing is tricky,
since that needs interrupts enabled and a wrongly timed interrupt
completely messes up the timing (and those functions are fairly slow,
too).

Ideally, you would use a hand-coded assembly loop with a known cycle
count for maximum accuracy, but that is of course not portable.
Fortunately, Arduino offers the pulseIn function which can take care of
exactly this. Since Arduino 1.6.5 this is fixed cycle count loop using
assembly, earlier versions used hardcoded cycle counting for the same
result.

This change removes the need for the `count` constructor argument,
though it is kept around for compatibility.